updated mapping functional updated dtos, requests, responses added new fields to entities removed authority classifier logic added authority question group to activity template config updated activity list extract logic (added object question info) updated question functional getting by types added functional to question (title, required) updated KnowledgeCategory entity structure updated structure converter functional created type converter functional created type converters (string, int, date) updated constants added functional to db initializer (parsing queries in new format) added init queries updated data, relationship and schema queries updated properties (added base servlet path) code refactoring

+++ src/main/java/com/datapath/checklistukraineapp/dto/QuestionDTO.java
@@ -11,7 +11,7 @@ public class QuestionDTO {
     private String name;
     private LocalDate dateCreated;
 
-    private Integer knowledgeCategoryId;
+    private Long knowledgeCategoryId;
     private String knowledgeCategoryValue;
     private String knowledgeCategoryTranslate;
